# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FondosEG API",
    "description": "API de Billetera Digital para aplicaciones externas. Autentica usando API Key y API Secret en los headers.",
    "version": "1.0.0",
    "contact": {
      "name": "Soporte FondosEG",
      "email": "soporte@FondosEG.com"
    }
  },
  "servers": [
    {
      "url": "https://tu-dominio.com/api/external",
      "description": "Servidor de producción"
    },
    {
      "url": "http://localhost:3000/api/external",
      "description": "Servidor de desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "ApiSecretAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-secret"
      }
    },
...
  },
  "paths": {
...
    "/wallet-transfer": {
      "post": {
        "summary": "Transferencia entre clientes (Billetera)",
        "description": "Transfiere dinero de una billetera de cliente a otra. Solo disponible para API keys con rol de cliente. No requiere verificación por código.",
        "security": [
          { "ApiKeyAuth": [] },
          { "ApiSecretAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["receiver_phone", "receiver_name", "amount"],
                "properties": {
                  "receiver_phone": { "type": "string", "description": "Teléfono del receptor" },
                  "receiver_name": { "type": "string", "description": "Nombre del receptor" },
                  "amount": { "type": "number", "description": "Monto a transferir" },
                  "currency": { "type": "string", "default": "XAF", "description": "Moneda" },
                  "notes": { "type": "string", "description": "Notas opcionales" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transferencia exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transfer_id": { "type": "string" },
                        "transfer_type": { "type": "string" },
                        "amount": { "type": "number" },
                        "currency": { "type": "string" },
                        "sender_name": { "type": "string" },
                        "receiver_name": { "type": "string" },
                        "receiver_phone": { "type": "string" },
                        "status": { "type": "string" },
                        "new_balance": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación o saldo insuficiente"
          },
          "401": {
            "description": "Autenticación fallida"
          },
          "403": {
            "description": "Permiso denegado - Solo clientes"
          }
        }
      }
    },
...
  }
}